:recycle: (MessagingApps) reorganize files
