+++ src/features/offer/components/OfferMessagingApps/OfferMessagingApps.tsx
@@ -1,10 +1,9 @@
 import React, { useCallback } from 'react'
 import { Social } from 'react-native-share'
 
+import { MessagingApps } from 'features/share/components/MessagingApps/MessagingApps'
 import { useShareOffer } from 'features/share/helpers/useShareOffer'
 import { analytics } from 'libs/analytics'
-
-import { MessagingApps } from './MessagingApps'
 
 type MessagingAppsProps = {
   isEvent: boolean
