+++ src/features/share/components/MessagingApps/MessagingApps.tsx
@@ -2,8 +2,8 @@ import React, { useCallback } from 'react'
 import { Social } from 'react-native-share'
 import styled from 'styled-components/native'
 
-import { InstalledMessagingApps } from 'features/offer/components/shareMessagingOffer/InstalledMessagingApps'
-import { MessagingAppContainer } from 'features/offer/components/shareMessagingOffer/MessagingAppContainer'
+import { InstalledMessagingApps } from 'features/share/components/MessagingApps/InstalledMessagingApps'
+import { MessagingAppContainer } from 'features/share/components/MessagingApps/MessagingAppContainer'
 import { WebShareModal } from 'features/share/pages/WebShareModal'
 import { ShareContent } from 'libs/share'
 import { useModal } from 'ui/components/modals/useModal'
