+++ src/features/venue/components/VenueMessagingApps/VenueMessagingApps.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react'
 import { Social } from 'react-native-share'
 
-import { MessagingApps } from 'features/offer/components/shareMessagingOffer/MessagingApps'
+import { MessagingApps } from 'features/share/components/MessagingApps/MessagingApps'
 import { useShareVenue } from 'features/share/helpers/useShareVenue'
 import { analytics } from 'libs/analytics'
